IT WORKS kinda
